switched database backend to mysql, moved all tables to same database, added user_id col to encounter and user tables

+++ internal/pkg/act/net_encoder.go
@@ -64,7 +64,7 @@ func writeTime(data *[]byte, value time.Time) {
 func EncodeEncounterBytes(value *Encounter) []byte {
 	data := make([]byte, 1)
 	data[0] = DataTypeEncounter
-	writeInt32(&data, value.ID)
+	writeInt32(&data, int32(value.ID))
 	writeTime(&data, value.StartTime)
 	writeTime(&data, value.EndTime)
 	writeString(&data, value.Zone)
@@ -78,7 +78,7 @@ func EncodeEncounterBytes(value *Encounter) []byte {
 func EncodeCombatantBytes(value *Combatant) []byte {
 	data := make([]byte, 1)
 	data[0] = DataTypeCombatant
-	writeInt32(&data, value.EncounterID)
+	writeInt32(&data, int32(value.EncounterID))
 	writeString(&data, value.Name)
 	writeString(&data, value.Job)
 	writeInt32(&data, value.Damage)
@@ -95,7 +95,7 @@ func EncodeCombatantBytes(value *Combatant) []byte {
 func EncodeLogLineBytes(value *LogLine) []byte {
 	data := make([]byte, 1)
 	data[0] = DataTypeLogLine
-	writeInt32(&data, value.EncounterID)
+	writeInt32(&data, int32(value.EncounterID))
 	writeTime(&data, value.Time)
 	writeString(&data, value.LogLine)
 	return data
